some more u8 to u16s thanks to deokishisu

+++ src/easy_chat.c
@@ -4988,7 +4988,7 @@ static void HideStartSelectButtons(void)
 
 static void TryAddInterviewObjectEvents(void)
 {
-    int graphicsId;
+    u16 graphicsId;
     u8 spriteId;
 
     switch (GetDisplayedPersonType())
